ARM SDT 2.11a / C++ 1.01.
2.11a's release date was around May 1997.

ARM C++ was available separately, but is built from the same source tree.

This was described to me as a snapshot from Lee Smith's laptop when he retired, rather than
a source release from a tagged repository. It has several duplicate files and a couple
where two differing versions exist in different directories.

Based on this, this may not be exactly 2.11a, though it fits the description in the
documentation closely and doesn't have the new features of 2.50.

I've applied this source to the repo on top of the 1998 MIPS port since it has many changes
that the MIPS 'mip' directory doesn't have, but it is more likely that the ARM
fork was worked on after the MIPS work was started.